refactor: clean up and move to own component
